requirements.txt
pyrogram==2.0.106
telethon==1.36.0
python-dotenv==1.0.1
aiofiles==23.2.1
pillow==10.0.0
uvloop==0.19.0
httpx==0.27.0
tgcrypto
pymongo==4.10.1
motor==3.5.1
